--- vms_fe/src/pages/DefaultPage.module.css
@@ -1,64 +1,0 @@
-.pageContainer {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: var(--black-400);
-  color: var(--white);
-}
-.nav {
-  width: 100%;
-}
-.mainContainer {
-  width: 100%;
-  max-width: 1200px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 0 24px;
-  padding-bottom: 133px;
-}
-.searchBarContainer {
-  width: 100%;
-  max-width: 1200px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 40px;
-  margin-bottom: 16px;
-}
-.listTitle {
-  font-size: 20px;
-  font-weight: 700;
-}
-.searchBarRight {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-.listContainer {
-  width: 100%;
-  max-width: 1200px;
-  margin-bottom: 40px;
-}
-@media (max-width: 744px) {
-  .searchBarContainer {
-    margin-top: 24px;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-  .listTitle {
-    font-size: 16px;
-    padding-bottom: 16px;
-  }
-  .searchBarRight {
-    width: 100%;
-    justify-content: space-between;
-  }
-  .listContainer {
-    overflow-x: auto;
-    padding-bottom: 12px;
-    margin-bottom: 0;
-  }
-}
