--- vms_fe/src/pages/LandingPage.module.css
@@ -1,360 +1,0 @@
-.container {
-  height: 100vh;
-  position: relative;
-}
-
-.scrollContainer {
-  height: 100vh;
-  overflow-y: scroll;
-  scroll-snap-type: y mandatory;
-  width: 100%;
-  position: relative;
-}
-
-.pageContainer {
-  height: 100vh;
-  scroll-snap-align: start;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-  box-sizing: border-box;
-}
-
-.sidebar {
-  position: fixed;
-  top: 50%;
-  right: 0;
-  transform: translateY(-50%);
-  width: 75px;
-  background-color: transparent;
-  color: #ffffff;
-  display: flex;
-  flex-direction: column;
-  padding: 20px;
-  z-index: 10;
-}
-
-.sidebarItem {
-  padding: 8px 10px;
-  margin-bottom: 8px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  border: 1px solid transparent;
-  font-size: 14px;
-  background-color: transparent;
-}
-
-.sidebarItem:first-child {
-  border-top: 1px solid #ffffff;
-}
-
-.sidebarItem:last-child {
-  border-bottom: 1px solid #ffffff;
-}
-
-.sidebarItem.active,
-.sidebarItem:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.backgroundGradient {
-  position: absolute;
-  inset: 0;
-  z-index: -1;
-}
-
-.backgroundGradient::before {
-  content: "";
-  position: absolute;
-  inset: 0;
-  background-image: url('assets/images/backgroundImge.jpg');
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
-  opacity: 0.15;
-  z-index: -1;
-}
-
-.introSection {
-  text-align: center;
-  margin-bottom: 64px;
-}
-
-.introTitle {
-  font-size: 40px;
-  font-weight: bold;
-  margin-bottom: 16px;
-  background: radial-gradient(ellipse at center, #b3b1b1 40%, #555353 60%);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-
-.introSubtitle {
-  font-size: 20px;
-  color: var(-webkit-background-clip);
-  margin-bottom: 32px;
-}
-
-.exploreButton {
-  background: linear-gradient(to right, #131313, #282828);
-  color: #fff;
-  font-weight: bold;
-  padding: 12px 24px;
-  border-radius: 9999px;
-  display: inline-flex;
-  align-items: center;
-  text-decoration: none;
-  transition: all 0.3s ease;
-  box-shadow: 0 0 15px rgba(255, 255, 255, 0.5);
-}
-
-.exploreButton:hover {
-  transform: scale(1.05);
-}
-
-.arrowIcon {
-  margin-left: 8px;
-}
-
-.chartSection {
-  display: grid;
-  grid-template-columns: 2.5fr 1fr;
-  gap: 48px;
-  width: 70%;
-  padding: 32px;
-  border-radius: 8px;
-  color: #ffffff;
-}
-
-.chartTitle {
-  font-size: 24px;
-  font-weight: bold;
-  margin-bottom: 16px;
-}
-
-.chartTextContainer {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  align-self: center;
-  gap: 16px;
-}
-
-.chartTextTitle {
-  font-size: 24px;
-  font-weight: 600;
-
-}
-
-.chartContainer {
-  height: 400px;
-  width: 100%;
-}
-
-.statsSection {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #2e2e2e;
-  padding: 32px;
-  border-radius: 8px;
-  color: #ffffff;
-  width: 70%;
-  height: 70%;
-}
-
-.statsCard {
-  text-align: center;
-  margin: 0 16px;
-}
-
-.statsIcon {
-  width: 40px;
-  height: 40px;
-  color: #ffffff;
-}
-
-.statsTitle {
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.statsValue {
-  font-size: 24px;
-  font-weight: bold;
-}
-
-.investmentPortfolio {
-  margin-top: 2rem;
-}
-
-.portfolioItems {
-  display: flex;
-  justify-content: space-around;
-}
-
-.portfolioItem {
-  background-color: #444;
-  padding: 20px;
-  border-radius: 10px;
-  text-align: center;
-  max-width: 300px;
-  margin: 10px;
-}
-
-.learnMoreButton {
-  display: inline-block;
-  margin-top: 10px;
-  padding: 5px 10px;
-  background-color: #0056b3;
-  color: #fff;
-  border-radius: 5px;
-  text-decoration: none;
-}
-
-.chartInsights {
-  margin-top: 2rem;
-  color: #ffffff;
-}
-
-.chartInsights h4 {
-  margin-bottom: 1rem;
-}
-
-.growthDetails {
-  margin-top: 2rem;
-}
-
-.growthChart,
-.growthCases {
-  margin-bottom: 1rem;
-}
-
-.investorSection {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #2e2e2e;
-  padding: 32px;
-  border-radius: 8px;
-  color: #ffffff;
-  width: 70%;
-  height: 80%;
-}
-
-.InvestorContainer {
-  min-height: 50vh;
-  background: inherit;
-  padding: 2rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-}
-
-.InvestorTitle {
-  font-size: 2.5rem;
-  font-weight: bold;
-  color: white;
-  margin-bottom: 2rem;
-}
-
-.InvestorDescription {
-  font-size: 1.25rem;
-  color: white;
-  margin-bottom: 3rem;
-  text-align: center;
-  max-width: 40rem;
-  white-space: normal;
-}
-
-.InvestorGrid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
-  gap: 1.5rem;
-  width: 100%;
-  max-width: 1200px;
-}
-
-.InvestorCard {
-  background-color: rgba(255, 255, 255, 0.2);
-  border-radius: 0.5rem;
-  padding: 1.0rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.InvestorImage {
-  width: 100px;
-  height: 100px;
-  border-radius: 50%;
-  margin-bottom: 1rem;
-}
-
-.InvestorName {
-  font-size: 1.25rem;
-  font-weight: semibold;
-  color: white;
-  margin-bottom: 0.5rem;
-}
-
-.InvestorRole {
-  font-size: 0.875rem;
-  color: rgba(255, 255, 255, 0.8);
-}
-
-.investmentTotal {
-  margin-top: 2rem;
-}
-
-.investmentChart {
-  height: 300px;
-  background-color: #444;
-}
-
-.startupsList {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.startupItem {
-  background-color: #444;
-  padding: 15px;
-  border-radius: 10px;
-  margin: 5px;
-}
-
-@media (max-width: 1200px) {
-  .chartSection {
-    display: flex;
-    flex-direction: column;
-    gap: 48px;
-    width: 70%;
-    padding: 32px;
-    border-radius: 8px;
-    color: #ffffff;
-  }
-}
-
-@media (max-width: 744px) {
-  .InvestorGrid {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .InvestorCard {
-    max-width: 80%;
-    margin-bottom: 1rem;
-  }
-
-
-}
